requirements.txt
ailia==1.3.0.6
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dlib==19.21.1
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.2.0
ftfy==6.2.0
h5py==3.11.0
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
lapx==0.5.9
lazy_loader==0.4
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-image==0.22.0
scipy==1.13.0
six==1.16.0
sympy==1.12
tifffile==2024.5.10
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.9.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
Werkzeug==3.0.3
zipp==3.18.1
